Rename subpackage positionprocessor into position
