+++ single-line-diagram/single-line-diagram-cgmes/single-line-diagram-cgmes-layout/src/main/java/com/powsybl/sld/cgmes/layout/LayoutToCgmesDlExporterTool.java
@@ -14,8 +14,8 @@ import com.powsybl.sld.AbstractSingleLineDiagramCommand;
 import com.powsybl.sld.cgmes.dl.conversion.CgmesDLExporter;
 import com.powsybl.sld.cgmes.dl.conversion.CgmesDLUtils;
 import com.powsybl.sld.layout.*;
-import com.powsybl.sld.layout.positionprocessor.positionbyclustering.PositionByClustering;
-import com.powsybl.sld.layout.positionprocessor.positionfromextension.PositionFromExtension;
+import com.powsybl.sld.layout.position.positionbyclustering.PositionByClustering;
+import com.powsybl.sld.layout.position.positionfromextension.PositionFromExtension;
 import com.powsybl.tools.Command;
 import com.powsybl.tools.Tool;
 import com.powsybl.tools.ToolOptions;
