+++ single-line-diagram/single-line-diagram-core/src/main/java/com/powsybl/sld/layout/position/BlockPositionner.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.sld.layout.positionprocessor;
+package com.powsybl.sld.layout.position;
 
 import com.powsybl.sld.model.cells.BusCell;
 import com.powsybl.sld.model.cells.InternCell;
