+++ single-line-diagram/single-line-diagram-core/src/main/java/com/powsybl/sld/layout/position/positionbyclustering/Links.java
@@ -4,10 +4,10 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.sld.layout.positionprocessor.positionbyclustering;
+package com.powsybl.sld.layout.position.positionbyclustering;
 
-import com.powsybl.sld.layout.positionprocessor.BSCluster;
-import com.powsybl.sld.layout.positionprocessor.HorizontalBusListManager;
+import com.powsybl.sld.layout.position.BSCluster;
+import com.powsybl.sld.layout.position.HorizontalBusListManager;
 import com.powsybl.sld.model.coordinate.Side;
 
 import java.util.*;
