+++ single-line-diagram/single-line-diagram-core/src/main/java/com/powsybl/sld/layout/position/positionbyclustering/PositionByClustering.java
@@ -4,13 +4,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.sld.layout.positionprocessor.positionbyclustering;
+package com.powsybl.sld.layout.position.positionbyclustering;
 
-import com.powsybl.sld.layout.positionprocessor.AbstractPositionFinder;
-import com.powsybl.sld.layout.positionprocessor.BSCluster;
-import com.powsybl.sld.layout.positionprocessor.HorizontalBusList;
-import com.powsybl.sld.layout.positionprocessor.Subsection;
-import com.powsybl.sld.layout.positionprocessor.VerticalBusSet;
+import com.powsybl.sld.layout.position.AbstractPositionFinder;
+import com.powsybl.sld.layout.position.BSCluster;
+import com.powsybl.sld.layout.position.HorizontalBusList;
+import com.powsybl.sld.layout.position.Subsection;
+import com.powsybl.sld.layout.position.VerticalBusSet;
 import com.powsybl.sld.model.cells.BusCell;
 import com.powsybl.sld.model.cells.Cell;
 import com.powsybl.sld.model.cells.ExternCell;
