+++ single-line-diagram/single-line-diagram-core/src/main/java/com/powsybl/sld/layout/position/positionfromextension/PositionFromExtension.java
@@ -4,11 +4,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.sld.layout.positionprocessor.positionfromextension;
+package com.powsybl.sld.layout.position.positionfromextension;
 
-import com.powsybl.sld.layout.positionprocessor.AbstractPositionFinder;
-import com.powsybl.sld.layout.positionprocessor.BSCluster;
-import com.powsybl.sld.layout.positionprocessor.VerticalBusSet;
+import com.powsybl.sld.layout.position.AbstractPositionFinder;
+import com.powsybl.sld.layout.position.BSCluster;
+import com.powsybl.sld.layout.position.VerticalBusSet;
 import com.powsybl.sld.model.cells.*;
 import com.powsybl.sld.model.coordinate.Side;
 import com.powsybl.sld.model.graphs.VoltageLevelGraph;
